Add files via upload
This system mainly uses DAQ system which is based on microcontroller
intel-8051, with a feature to save the data in a file for back up. In it the interfacing is
done by using I2C,SPI and UART protocols. This project is very useful in many data
logging applications to detect the various parameters.